handle edge case with road_building
it's not possible for a player with 14 or 15 roads already built to get any more out of a road building card. to prevent the user from getting trapped in a state with no possible actions, we have to check before we lock ourselves into the road building state.

interestingly, it might be possible to _still_ get caught in a bind if more roads are available, but the active player is totally boxed in and has no legal spots on which to build roads. not yet sure how i should handle that.